Support to enable/disable the RTC integration per project
It can now be configured per project whether the RTC integration is
enabled or not. To enable the RTC integration for a project the
project must have the following entry in its `project.config` file in
the `refs/meta/config` branch:

```
  [plugin "rtc"]
    enabled = true
```

If `plugin.rtc.enabled` is not specified in the `project.config` file
the value is inherited from the parent project. If it is also not set
on any parent project the RTC integration is disabled for this
project.

By setting `plugin.rtc.enabled` to true in the `project.config` of the
`All-Projects` project the RTC integration can be enabled by default
for all projects. During the initialization of the plugin you are asked
if the RTC integration should be enabled by default for all projects
and if yes this setting in the `project.config` of the `All-Projects`
project is done automatically.``

This uses new functionality in Gerrit and hence the Gerrit API
version needs to be changed to 2.9-SNAPSHOT.`

Change-Id: If24cb7f9cf9f62aff9eeb376aff0e6a4724ed020
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>